+++ packages/agent/test/utils/query-parser/body-parser.test.ts
@@ -1,8 +1,8 @@
 import { CollectionSchema } from '@forestadmin/datasource-toolkit';
 import { createMockContext } from '@shopify/jest-koa-mocks';
 
+import * as factories from '../../__factories__';
 import BodyParser from '../../src/utils/body-parser';
-import * as factories from '../__factories__';
 
 describe('BodyParser', () => {
   describe('parseSelectionIds', () => {
